+++ asana/resources/attachments.py
@@ -1,5 +1,5 @@
 
-from ._attachments import _Attachments
+from .gen.attachments import _Attachments
 
 class Attachments(_Attachments):
     """Attachments resource"""
