+++ asana/resources/gen/stories.py
