+++ asana/resources/projects.py
@@ -1,5 +1,5 @@
 
-from ._projects import _Projects
+from .gen.projects import _Projects
 
 class Projects(_Projects):
     """Projects resource"""
