+++ asana/resources/stories.py
@@ -1,5 +1,5 @@
 
-from ._stories import _Stories
+from .gen.stories import _Stories
 
 class Stories(_Stories):
     """Stories resource"""
