+++ asana/resources/tasks.py
@@ -1,5 +1,5 @@
 
-from ._tasks import _Tasks
+from .gen.tasks import _Tasks
 
 class Tasks(_Tasks):
     """Tasks resource"""
