+++ asana/resources/teams.py
@@ -1,5 +1,5 @@
 
-from ._teams import _Teams
+from .gen.teams import _Teams
 
 class Teams(_Teams):
     """Teams resource"""
