+++ asana/resources/users.py
@@ -1,5 +1,5 @@
 
-from ._users import _Users
+from .gen.users import _Users
 
 class Users(_Users):
     """Users resource"""
