+++ asana/resources/workspaces.py
@@ -1,5 +1,5 @@
 
-from ._workspaces import _Workspaces
+from .gen.workspaces import _Workspaces
 
 class Workspaces(_Workspaces):
     """Workspaces resource"""
